requirements.txt
decorator
Django
django-extensions
djangorestframework
Faker
ipdb
ipython
ipython-genutils
jedi
parso
pexpect
pickleshare
prompt-toolkit
ptyprocess
Pygments
python-dateutil
pytz
simplegeneric
six
text-unidecode
traitlets
wcwidth